libimage: Add ExecStdout helper and FormatAs helper
This very simple set of helper routines makes it very easy for us to format
a path with a given filesystem using a known set of sane arguments, and
ensures we don't try to do anything we don't know how to do, i.e. random
filesystem types from user input.

The new ExecStdout helper will likely become part of the larger library in
general, because it's kinda mundane executing commands in golang.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>